requirements.txt
litellm=1.67.2
google-adk=0.3.0
azure-kusto-data
pythonnet